requirements.txt
python-telegram-bot[webhooks]
python-dotenv
groq
scikit-learn
pandas
numpy
joblib
langdetect